I tried..

+++ src/main/java/com/revivalmodding/revivalcore/core/common/tileentity/IProcessCraftSystem.java
@@ -23,7 +23,7 @@ public interface IProcessCraftSystem<R extends RVRecipe> extends ICraftSystem<R>
 	// For gui rendering
 	default float getProgressionStage()
 	{
-		float f = this.getProcessTimer() / this.getRecipe().getCraftTime();
+		float f = this.getProcessTimer() / (float)this.getRecipe().getCraftTime();
 		return f;
 	}
